requirements.txt
openai
markdown
streamlit
python-dotenv
pandas